requirements.txt
pandas
numpy
matplotlib
scikit-learn
scipy
tensorflow
keras
biosppy
neurokit2
pyserial
joblib